backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.12
frozenlist==1.6.2
google-api-core==2.25.0
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.63
langchain-openai==0.3.19
langchain-pinecone==0.2.6
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.44
multidict==6.4.4
numpy==2.2.6
openai==1.84.0
orjson==3.10.18
packaging==24.2
pinecone==6.0.2
pinecone-plugin-assistant==1.6.1
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
syrupy==4.9.1
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
yarl==1.20.0
zstandard==0.23.0
